Add files via upload
Hi Lukas,

Sorroy about the mistakes. Here is the new design of the Si+heaters chip. I fix the optial IO error. The 2 and the 4 GC is used as the optical input and the 1 and the 3 GC are used as the optical output. 

There are still two errors, overlapping component, occurs at the connection of a stright waveguide and the pins of the spiral. I suppose it can be ignored.

Thank you.
Yiran Guan