Add files via upload
Hi Lukas,

Here is the new design of the Si+heaters chip. According to the rules. GC 1 and 2 are used as the output and input of device 1, respectively. GC 3 and 4 are used as the output and input for device 2, respectively.

There are still two errors, overlapping component, occurs at the connection of a stright waveguide and the pins of the spiral. I suppose it can be ignored.

Please check my design again. Thanks.
Yiran Guan